+++ CHANGELOG.md
@@ -7,3 +7,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 The contents of this changelog are focused on changes to overall behaviour of this library, as relevant to a developer using its classes. Internal changes that don't affect this behaviour are ommitted, or simply noted as 'Minor internal changes', 'Minor performance optimizations', etc.
 
 ## [UNRELEASED]
+### Added
+ - `PclExceptionWithCode::class` - Extension of `\Exception` to provide predefined 'exception codes' for debugging/error handling purposes